Add localStorage support to valuesGetProxy and deleteValue

+++ src/values.ts
@@ -187,7 +187,7 @@ export function valuesProxy<Values extends string>(
 }
 
 /**
- * Requires the `GM.getValue` grant.
+ * Requires the `GM.getValue` grant or falls back to using localStorage.
  * Get a Proxy that wraps `GM.getValue` for better typing.
  * Useful when a value may be modified by multiple different sources,
  * meaning the value will need to be retrieved from GM every time.
@@ -211,17 +211,27 @@ export function valuesProxy<Values extends string>(
 export function valuesGetProxy<Values extends string>(
   values: ValuesObject<Values>
 ): ValuesPromiseObject<Values> {
+  const grants = checkGrants('getValue')
+
   /** Handle gets to the values object */
   const handler: ProxyHandler<ValuesObject<Values>> = {
     get(target, prop: Values): Promise<GM.Value> {
       return new Promise((resolve, reject) => {
         // Check if the property is a part of the passed values
         if (prop in target) {
-          GM.getValue(prop).then(value => {
-            // Resolve with the value if it's defined
-            if (value !== undefined) resolve(value)
+          // Using GreaseMonkey
+          if (grants) {
+            GM.getValue(prop).then(value => {
+              // Resolve with the value if it's defined
+              if (value !== undefined) resolve(value)
+              else reject()
+            })
+            // Using localStorage
+          } else {
+            const value = localStorage.getItem(prop)
+            if (value !== null) resolve(value)
             else reject()
-          })
+          }
         } else {
           reject()
         }
@@ -238,7 +248,7 @@ export function valuesGetProxy<Values extends string>(
 }
 
 /**
- * Requires the `GM.deleteValue` grant.
+ * Requires the `GM.deleteValue` grant or falls back to localStorage.
  * Deletes a value from a values object.
  * This is only useful if you're using TypeScript or your editor has typing support.
  * If that doesn't describe your use case, then use `GM.deleteValue` instead.
@@ -254,7 +264,11 @@ export function deleteValue<Values extends string, ToDelete extends Values>(
 ): Promise<Omit<ValuesObject<Values>, ToDelete>> {
   return new Promise(async (resolve, reject) => {
     if (toDelete in values) {
-      await GM.deleteValue(toDelete)
+      // Using GreaseMonkey
+      if (checkGrants('deleteValue')) await GM.deleteValue(toDelete)
+      // Using localStorage
+      else localStorage.removeItem(toDelete)
+
       delete values[toDelete]
       resolve(values)
     }
